Remove code complexity and take a 200 bytes penalty
Now that we only keep the diff, there is no need to compress duplicated
zeroes. It only saves about 200 bytes in the base frame and almost
nothing in the diffs. Remove the code to reduce code complexity,
especially as I'll have to unpack the bytes in C.

Goes from ~910 bytes to ~1100 bytes. Still a considerable optimization
compared to the original ~5000 bytes.