Generate more compact code.
Flatten the frames into one array for all the frames bytes and one array
for all the frames regions.
This forces us to keep track of the boundaries in other arrays, but it
means that a frame can be regenerated dynamically from an index, so I
will be able to write C code to work with these arrays without using
macros.